django-backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.4.0
django-cors-headers==4.4.0
redis==5.0.8